Add Hero tests
