+++ apps/climatemappedafrica/src/components/Hero/Hero.snap.js
@@ -1,0 +1,103 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`Hero Component renders unchanged 1`] = `
+<div>
+  <div
+    class="makeStyles-root-1"
+  >
+    <div
+      class="makeStyles-background-2"
+    >
+      <img
+        data-nimg="fill"
+        decoding="async"
+        loading="lazy"
+        src="/url"
+        style="position: absolute; height: 100%; width: 100%; left: 0px; top: 0px; right: 0px; bottom: 0px; color: transparent;"
+      />
+    </div>
+    <div
+      class="makeStyles-root-14 makeStyles-root-12 makeStyles-root-10 makeStyles-section-4 makeStyles-fixed-11 MuiBox-root css-0"
+    >
+      <div
+        class="MuiGrid-root MuiGrid-container css-11lq3yg-MuiGrid-root"
+      >
+        <div
+          class="MuiBox-root css-h9r63z"
+        >
+          <div
+            class="MuiGrid-root MuiGrid-item MuiGrid-grid-lg-1 css-9jo7xm-MuiGrid-root"
+          />
+        </div>
+        <div
+          class="MuiGrid-root MuiGrid-item MuiGrid-grid-xs-12 MuiGrid-grid-md-7 MuiGrid-grid-lg-6 css-1totwg2-MuiGrid-root"
+        >
+          <header
+            class="MuiBox-root css-0"
+            classes="[object Object]"
+          >
+            <div
+              class="MuiBox-root css-1ki7zn5"
+            >
+              <h1
+                class="MuiTypography-root MuiTypography-h1 css-qt60le-MuiTypography-root"
+              >
+                Data to hold your government accountable
+              </h1>
+            </div>
+            <div
+              class="MuiBox-root css-1xaekgw"
+            >
+              <h6
+                class="MuiTypography-root MuiTypography-subtitle1 css-1d1ipbv-MuiTypography-root"
+              >
+                PesaYetu helps journalists, researchers and activists transform their work with in-depth county-specific information. Get started now with datasets from Kenya.
+
+              </h6>
+            </div>
+          </header>
+          <div
+            class="makeStyles-root-16"
+            id="location-search"
+          >
+            <p
+              class="MuiTypography-root MuiTypography-body1 makeStyles-label-19 makeStyles-dropdownTitle-8 css-6hq4le-MuiTypography-root"
+            >
+              Search for a location
+            </p>
+            <div
+              blocktype="hero"
+              boundary="[object Object]"
+              class="MuiInputBase-root makeStyles-inputRoot-17 MuiInputBase-colorPrimary css-1illudf-MuiInputBase-root"
+              slug="hero"
+            >
+              <input
+                aria-label="search"
+                class="MuiInputBase-input makeStyles-input-21 css-yz9k0d-MuiInputBase-input"
+                id="670e3996766697e7feb349d5"
+                placeholder="Search for a location"
+                type="text"
+                value=""
+              />
+            </div>
+            <div
+              class="makeStyles-suggestions-22"
+            />
+          </div>
+        </div>
+        <div
+          class="MuiBox-root css-h9r63z"
+        >
+          <div
+            class="MuiGrid-root MuiGrid-item MuiGrid-grid-md-5 css-1f9gci5-MuiGrid-root"
+          >
+            <div
+              class="MuiBox-root css-4tnojn"
+            />
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+`;
