+++ apps/climatemappedafrica/src/components/Hero/Map.js
